requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
babel==2.18.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
courlan==1.3.2
cryptography==46.0.5
dateparser==1.3.0
distro==1.9.0
duckduckgo_search==8.1.1
filelock==3.24.3
fsspec==2026.2.0
griffe==1.15.0
h11==0.16.0
hf-xet==1.2.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.4
MarkupSafe==3.0.3
mcp==1.26.0
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
openai==2.26.0
openai-agents==0.10.5
packaging==26.0
pillow==12.1.1
primp==1.1.2
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.1.2
six==1.17.0
sniffio==1.3.1
sse-starlette==3.3.2
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tld==0.13.2
tokenizers==0.22.2
torch==2.2.2
tqdm==4.67.3
trafilatura==2.0.0
transformers==4.57.6
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
